Add permissions to standard activities

+++ app/Http/Controllers/ActivityDivisionController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use App\ActivityDivision;
+use Illuminate\Contracts\Auth\Guard;
 use Illuminate\Http\Request;
 
 class ActivityDivisionController extends Controller
@@ -12,17 +13,31 @@ class ActivityDivisionController extends Controller
 
     public function index()
     {
+        if (\Gate::denies('read', 'std-activities')) {
+            flash("You don't have access to this page");
+            return \Redirect::to('/');
+        }
         $activityDivisions = ActivityDivision::tree()->appendActivity()->paginate(25);
         return view('activity-division.index', compact('activityDivisions'));
     }
 
     public function create()
     {
+        if (\Gate::denies('write', 'std-activities')) {
+            flash("You don't have access to this page");
+            return \Redirect::to('/');
+        }
+
         return view('activity-division.create');
     }
 
     public function store(Request $request)
     {
+        if (\Gate::denies('write', 'std-activities')) {
+            flash("You don't have access to this page");
+            return \Redirect::to('/');
+        }
+
         $this->validate($request, $this->rules);
 
         ActivityDivision::create($request->all());
@@ -34,16 +49,31 @@ class ActivityDivisionController extends Controller
 
     public function show(ActivityDivision $activity_division)
     {
+        if (\Gate::denies('read', 'std-activities')) {
+            flash("You don't have access to this page");
+            return \Redirect::to('/');
+        }
+
         return view('activity-division.show', compact('activity_division'));
     }
 
     public function edit(ActivityDivision $activity_division)
     {
+        if (\Gate::denies('write', 'std-activities')) {
+            flash("You don't have access to this page");
+            return \Redirect::to('/');
+        }
+
         return view('activity-division.edit', compact('activity_division'));
     }
 
     public function update(ActivityDivision $activity_division, Request $request)
     {
+        if (\Gate::denies('write', 'std-activities')) {
+            flash("You don't have access to this page");
+            return \Redirect::to('/');
+        }
+
         $this->validate($request, $this->rules);
 
         $activity_division->update($request->all());
@@ -55,6 +85,11 @@ class ActivityDivisionController extends Controller
 
     public function destroy(ActivityDivision $activity_division)
     {
+        if (\Gate::denies('delete', 'std-activities')) {
+            flash("You don't have access to this page");
+            return \Redirect::to('/');
+        }
+
         $activity_division->delete();
 
         flash('Activity division has been deleted', 'success');
@@ -64,8 +99,11 @@ class ActivityDivisionController extends Controller
 
     public function import()
     {
+        if (\Gate::denies('write', 'std-activities')) {
+            flash("You don't have access to this page");
+            return \Redirect::to('/');
+        }
 
-        //ActivityDivision::truncate();
         $path = storage_path('files/division.csv');
         $handle = fopen($path, "r");
 
@@ -95,5 +133,4 @@ class ActivityDivisionController extends Controller
         fclose($handle);
         return \Redirect::route('activity-division.index');
     }
-
 }
