+++ app/Http/Controllers/StdActivityController.php
@@ -17,6 +17,11 @@ class StdActivityController extends Controller
 
     public function index()
     {
+        if (\Gate::denies('read', 'std-activities')) {
+            flash("You don't have access to this page");
+            return \Redirect::to('/');
+        }
+
         $filter = new StdActivityFilter(StdActivity::query(), session('filters.std-activity'));
         $stdActivities = $filter->filter()->paginate(100);
 
@@ -25,11 +30,21 @@ class StdActivityController extends Controller
 
     public function create()
     {
+        if (\Gate::denies('write', 'std-activities')) {
+            flash("You don't have access to this page");
+            return \Redirect::to('/');
+        }
+
         return view('std-activity.create');
     }
 
     public function store(Request $request)
     {
+        if (\Gate::denies('write', 'std-activities')) {
+            flash("You don't have access to this page");
+            return \Redirect::to('/');
+        }
+
         $this->validate($request, $this->rules);
 
         $activity = StdActivity::create($request->all());
@@ -42,16 +57,31 @@ class StdActivityController extends Controller
 
     public function show(StdActivity $std_activity)
     {
+        if (\Gate::denies('read', 'std-activities')) {
+            flash("You don't have access to this page");
+            return \Redirect::to('/');
+        }
+
         return view('std-activity.show', compact('std_activity'));
     }
 
     public function edit(StdActivity $std_activity)
     {
+        if (\Gate::denies('write', 'std-activities')) {
+            flash("You don't have access to this page");
+            return \Redirect::to('/');
+        }
+
         return view('std-activity.edit', compact('std_activity'));
     }
 
     public function update(StdActivity $std_activity, Request $request)
     {
+        if (\Gate::denies('write', 'std-activities')) {
+            flash("You don't have access to this page");
+            return \Redirect::to('/');
+        }
+
         $this->validate($request, $this->rules);
 
         $std_activity->update($request->all());
@@ -64,6 +94,11 @@ class StdActivityController extends Controller
 
     public function destroy(StdActivity $std_activity)
     {
+        if (\Gate::denies('delete', 'std-activities')) {
+            flash("You don't have access to this page");
+            return \Redirect::to('/');
+        }
+
         $std_activity->delete();
 
         flash('Std activity has been deleted', 'success');
@@ -73,11 +108,21 @@ class StdActivityController extends Controller
 
     function import()
     {
+        if (\Gate::denies('write', 'std-activities')) {
+            flash("You don't have access to this page");
+            return \Redirect::to('/');
+        }
+
         return view('std-activity.import');
     }
 
     function postImport(Request $request)
     {
+        if (\Gate::denies('write', 'std-activities')) {
+            flash("You don't have access to this page");
+            return \Redirect::to('/');
+        }
+
         $this->validate($request, [
             'file' => 'required|file'//|mimes:xls,xlsx'
         ]);
@@ -95,6 +140,11 @@ class StdActivityController extends Controller
 
     function filters(Request $request)
     {
+        if (\Gate::denies('read', 'std-activities')) {
+            flash("You don't have access to this page");
+            return \Redirect::to('/');
+        }
+
         $data = $request->only(['name', 'division_id']);
         \Session::set('filters.std-activity', $data);
 
@@ -115,20 +165,34 @@ class StdActivityController extends Controller
 
     function exportAllActivities()
     {
+        if (\Gate::denies('read', 'std-activities')) {
+            flash("You don't have access to this page");
+            return \Redirect::to('/');
+        }
+
         $this->dispatch(new ExportStdActivitiesJob());
     }
 
     function modifyAllActivities()
     {
+        if (\Gate::denies('write', 'std-activities')) {
+            flash("You don't have access to this page");
+            return \Redirect::to('/');
+        }
+
         return view('std-activity.modify');
     }
 
     function postModifyAllActivities(Request $request)
     {
+        if (\Gate::denies('write', 'std-activities')) {
+            flash("You don't have access to this page");
+            return \Redirect::to('/');
+        }
+
         $file = $request->file('file');
         $this->dispatch(new ModifyPublicStdActivitiesJob($file));
 
         return redirect()->back();
-
     }
 }
